reinstal NM on failure update

+++ src/helpers/utility.ts
@@ -292,9 +292,11 @@ let utiltiy = {
     fse.readdir(nmSwapFileDir, function(err, files) {
         if (err) {
           // if there is no folder reinstall
-          fse.unlink(nmInstallationSuccess, err => {
-            if (err) logger.error(err);
-          });
+          if (err.code === 'ENOENT'){
+             fse.unlink(nmInstallationSuccess, err => {
+                if (err) logger.error(err);
+              }); 
+          }
           return false;
         } else {
            if (!files.length) {
